# docs/swagger.yaml
basePath: /
definitions:
  domain.CreatePostRequest:
    properties:
      category_ids:
        items:
          type: string
        type: array
      content:
        type: string
      cover:
        type: string
      subTitle:
        type: string
      summary:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - content
    - title
    type: object
  domain.CreatePostResponse:
    properties:
      id:
        type: string
    type: object
  domain.CreateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  domain.ListTagsResponse:
    properties:
      tags:
        items:
          $ref: '#/definitions/domain.Tag'
        type: array
    type: object
  domain.LoginRequest:
    properties:
      password:
        type: string
      user:
        type: string
    required:
    - password
    - user
    type: object
  domain.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  domain.Post:
    properties:
      authorID:
        type: string
      category_id:
        items:
          type: string
        type: array
      content:
        type: string
      contentHtml:
        type: string
      cover:
        type: string
      draft:
        type: boolean
      id:
        type: string
      md5:
        type: string
      postId:
        type: string
      subTitle:
        type: string
      summary:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  domain.SignupResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  domain.Tag:
    properties:
      id:
        type: string
      name:
        type: string
      tagId:
        type: string
    type: object
host: 127.0.0.1:8081
info:
  contact: {}
  description: IvanApi Service
  title: go_zen blog system API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 注册用户，
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: 用户登录
        in: body
        name: object
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LoginResponse'
      summary: 用户登录接口
      tags:
      - 用户相关接口
  /posts/:id:
    get:
      consumes:
      - application/json
      description: 根据id获取文章，如果为空返回错误
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: 文章id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
      summary: 根据id获取文章
      tags:
      - 文章相关接口
  /posts/create:
    post:
      consumes:
      - application/json
      description: 创建文章，返回文章Id
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建文章的参数
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/domain.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CreatePostResponse'
      security:
      - Authorization: []
      summary: 创建文章
      tags:
      - 文章相关接口
  /posts/upload:
    post:
      consumes:
      - multipart/form-data
      description: 从文件上传创建文章，返回文章Id
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要上传的markdown文件
        in: formData
        name: mdFile
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CreatePostResponse'
      summary: 从文件上传创建文章
      tags:
      - 文章相关接口
  /signup:
    post:
      consumes:
      - application/json
      description: 注册用户，
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - in: query
        name: email
        required: true
        type: string
      - in: query
        name: name
        required: true
        type: string
      - in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SignupResponse'
      summary: 注册用户
      tags:
      - 用户相关接口
  /tag/create:
    post:
      consumes:
      - application/json
      description: 创建Tag，返回Tag
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: tagName
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag
          schema:
            $ref: '#/definitions/domain.Tag'
      summary: 创建Tag
      tags:
      - Tag相关接口
  /tag/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag
          schema:
            $ref: '#/definitions/domain.ListTagsResponse'
      summary: 获取所有tag
      tags:
      - Tag相关接口
  /tag/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 根据Id 删除tag
      tags:
      - Tag相关接口
  /tag/update:
    post:
      consumes:
      - application/json
      description: 创建Tag，返回Tag
      parameters:
      - description: Request ID
        in: header
        name: request_id
        required: true
        type: string
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: tag
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/domain.Tag'
      produces:
      - application/json
      responses:
        "200":
          description: Tag
          schema:
            $ref: '#/definitions/domain.Tag'
      summary: 更新Tag
      tags:
      - Tag相关接口
swagger: "2.0"